+++ src/indicators/ema.rs
@@ -53,6 +53,13 @@ impl IndicatorConfig for EmaConfig {
     fn convergence(&self) -> usize {
         self.length
     }
+
+    fn to_builder(&self) -> Self::Builder {
+        EmaConfigBuilder {
+            length: Some(self.length),
+            source: self.source,
+        }
+    }
 }
 
 impl EmaConfig {
@@ -518,6 +525,12 @@ mod tests {
 
             assert!(set.contains(&b));
             assert!(!set.contains(&c));
+        }
+
+        #[test]
+        fn to_builder_roundtrip() {
+            let config = EmaConfig::hl2(nz(12));
+            assert_eq!(config.to_builder().build(), config);
         }
     }